+++ src/constants/routes.js
@@ -4,4 +4,4 @@ export const FILE_DOWNLOAD = `${API}/files/:fileId/download`;
 export const FOLDER_LIST_ITEMS = `${API}/folders/:folderId/items`;
 export const FOLDER_CREATE = `${API}/folders/create`;
 export const FOLDER_THUMBNAILS = `${API}/folders/:folderId/thumbnails`;
-export const FILES_CHECKOUT = `${API}/files/checkout`;
+export const FOLDER_FIND = `${API}/folders/:folderPath`;